backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djoser==2.2.0
drf-base64==2.0
fpdf==1.7.2
gunicorn==20.1.0
isort==5.12.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
python-dotenv==1.0.0
reportlab==3.6.12
sqlparse==0.4.4
